requirements.txt
transformers 
peft 
accelerate 
datasets 
sentence-transformers 
google-generativeai 
fastapi 
uvicorn
fitz